Made files for each class.
