+++ src/main/java/Calculator.java
@@ -1,0 +1,9 @@
+class Calculator {
+    String names = "";
+    double price;
+
+    void calculate(String inputName, double inputPrice) {
+        names = names + "\n" + "\"" + inputName + "\"";
+        price += inputPrice;
+    }
+}
