+++ src/main/java/Main.java
@@ -44,53 +44,3 @@ class Main {
     }
 }
 
-class Calculator {
-    String names = "";
-    double price;
-
-    void calculate(String inputName, double inputPrice) {
-        names = names + "\n" + "\"" + inputName + "\"";
-        price += inputPrice;
-    }
-}
-
-class EndFormater {
-    String rubles = "руб.";
-
-    String formatRubleEnding(double result) {
-        double roundedEnding = Math.floor(result);
-        double lefts = roundedEnding % 100;
-        double lastDigit = lefts % 10;
-        if (((lefts >= 5) && (lefts <= 20)) || (lefts == 0) || ((lastDigit >= 5) && (lastDigit <= 9))) {
-            return rubles = "рублей";
-        } else {
-            if ((lefts == 2) || (lefts == 3) || (lefts == 4) || (lastDigit == 2) || (lastDigit == 3) || (lastDigit == 4)) {
-                return rubles = "рубля";
-            } else if (((lefts == 1)) || (lastDigit == 1)) {
-                return rubles = "рубль";
-            }
-            return rubles;
-        }
-    }
-}
-
-class CountPeople {
-    Scanner scanner = new Scanner(System.in);
-
-    void typeErrorMessage() {
-        System.out.println("Это некорректное значение для подсчёта. Попробуйте еще раз.");
-    }
-
-    int countPerson(int input) {
-        do {
-            System.out.println("На скольких человек необходимо разделить счёт? " + "Введите положительное целое число, больше 1.");
-            while (!scanner.hasNextInt()) {
-                typeErrorMessage();
-                scanner.next();
-            }
-            input = scanner.nextInt();
-        } while (input <= 0);
-        return input;
-    }
-}
-
